+++ core/com.b2international.snowowl.core/src/com/b2international/snowowl/core/codesystem/version/CodeSystemVersionCreateRequest.java
@@ -17,15 +17,18 @@ package com.b2international.snowowl.core.codesystem.version;
 
 import static com.b2international.snowowl.core.internal.locks.DatastoreLockContextDescriptions.CREATE_VERSION;
 
-import java.time.Instant;
+import java.time.LocalDate;
 import java.util.Collection;
-import java.util.Date;
 import java.util.List;
 import java.util.Map;
 import java.util.Map.Entry;
+import java.util.Objects;
+import java.util.Optional;
 import java.util.concurrent.TimeUnit;
 import java.util.function.Function;
 import java.util.stream.Collectors;
+
+import javax.validation.constraints.NotNull;
 
 import org.eclipse.core.runtime.IProgressMonitor;
 import org.eclipse.core.runtime.SubMonitor;
@@ -42,9 +45,8 @@ import com.b2international.snowowl.core.authorization.AccessControl;
 import com.b2international.snowowl.core.branch.Branch;
 import com.b2international.snowowl.core.codesystem.CodeSystem;
 import com.b2international.snowowl.core.codesystem.CodeSystemRequests;
+import com.b2international.snowowl.core.codesystem.CodeSystemVersion;
 import com.b2international.snowowl.core.codesystem.CodeSystemVersionEntry;
-import com.b2international.snowowl.core.date.DateFormats;
-import com.b2international.snowowl.core.date.EffectiveTimes;
 import com.b2international.snowowl.core.domain.exceptions.CodeSystemNotFoundException;
 import com.b2international.snowowl.core.events.Request;
 import com.b2international.snowowl.core.identity.Permission;
@@ -57,6 +59,7 @@ import com.b2international.snowowl.core.request.BranchRequest;
 import com.b2international.snowowl.core.request.CommitResult;
 import com.b2international.snowowl.core.request.RepositoryRequest;
 import com.b2international.snowowl.core.request.RevisionIndexReadRequest;
+import com.b2international.snowowl.core.request.SearchResourceRequest.SortField;
 import com.b2international.snowowl.core.terminology.TerminologyRegistry;
 import com.b2international.snowowl.eventbus.IEventBus;
 import com.fasterxml.jackson.annotation.JsonProperty;
@@ -78,18 +81,20 @@ final class CodeSystemVersionCreateRequest implements Request<ServiceProvider, B
 	@JsonProperty
 	String description;
 	
-	@NotEmpty
+	@NotNull
 	@JsonProperty
-	String effectiveTime;
+	LocalDate effectiveTime;
 	
 	@NotEmpty
 	@JsonProperty
 	String codeSystemShortName;
 	
+	@JsonProperty
+	boolean force;
+	
 	// local execution variables
 	private transient Multimap<DatastoreLockContext, DatastoreLockTarget> lockTargetsByContext;
 	private transient Map<String, CodeSystem> codeSystemsByShortName;
-	private transient Date effectiveTimeDate;
 	
 	@Override
 	public Boolean execute(ServiceProvider context) {
@@ -104,33 +109,41 @@ final class CodeSystemVersionCreateRequest implements Request<ServiceProvider, B
 			throw new CodeSystemNotFoundException(codeSystemShortName);
 		}
 		
-		// check that the new versionId does not conflict with any other currently available branch
-		final String newVersionPath = String.join(Branch.SEPARATOR, codeSystem.getBranchPath(), versionId);
-		final String repositoryId = codeSystem.getRepositoryId();
-		
 		// validate new path
 		RevisionBranch.BranchNameValidator.DEFAULT.checkName(versionId);
 		
-		try {
-			Branch branch = RepositoryRequests.branching()
-				.prepareGet(newVersionPath)
-				.build(repositoryId)
-				.execute(context.service(IEventBus.class))
-				.getSync(1, TimeUnit.MINUTES);
-			// allow deleted version branches to be reused for versioning
-			if (!branch.isDeleted()) {
-				throw new ConflictException("An existing branch with path '%s' conflicts with the specified version identifier.", newVersionPath);
-			}
-		} catch (NotFoundException e) {
-			// ignore
-		}
-		
-		this.effectiveTimeDate = EffectiveTimes.parse(effectiveTime, DateFormats.SHORT);
-		
 		final List<CodeSystem> codeSystemsToVersion = codeSystem.getDependenciesAndSelf()
-			.stream()
-			.map(codeSystemsByShortName::get)
-			.collect(Collectors.toList());
+				.stream()
+				.map(codeSystemsByShortName::get)
+				.collect(Collectors.toList());
+		
+		for (CodeSystem cs : codeSystemsToVersion) {
+			// check that the new versionId does not conflict with any other currently available branch
+			final String newVersionPath = String.join(Branch.SEPARATOR, cs.getBranchPath(), versionId);
+			final String repositoryId = cs.getRepositoryId();
+			
+			if (!force) {
+				// branch needs checking in the non-force cases only
+				try {
+					
+					Branch branch = RepositoryRequests.branching()
+						.prepareGet(newVersionPath)
+						.build(repositoryId)
+						.execute(context.service(IEventBus.class))
+						.getSync(1, TimeUnit.MINUTES);
+					
+					if (!branch.isDeleted()) {
+						throw new ConflictException("An existing branch with path '%s' conflicts with the specified version identifier.", newVersionPath);
+					}
+
+				} catch (NotFoundException e) {
+					// branch does not exist, ignore
+				}
+			}
+			
+			// if there is no conflict, delete the branch (the request also ignores non-existent branches)
+			deleteBranch(context, newVersionPath, repositoryId);
+		}
 		
 		acquireLocks(context, user, codeSystemsToVersion);
 		
@@ -139,13 +152,13 @@ final class CodeSystemVersionCreateRequest implements Request<ServiceProvider, B
 			
 			codeSystemsToVersion.forEach(codeSystemToVersion -> {
 				// check that the specified effective time is valid in this code system
-				validateEffectiveTime(context, codeSystem);
-				new RepositoryRequest<>(repositoryId,
+				validateVersion(context, codeSystem);
+				new RepositoryRequest<>(codeSystemToVersion.getRepositoryId(),
 					new BranchRequest<>(codeSystem.getBranchPath(),
 						new RevisionIndexReadRequest<CommitResult>(
 							context.service(RepositoryManager.class).get(codeSystemToVersion.getRepositoryId())
 								.service(VersioningRequestBuilder.class)
-								.build(new VersioningConfiguration(user, codeSystemToVersion.getShortName(), versionId, description, EffectiveTimes.getEffectiveTime(effectiveTimeDate)))
+								.build(new VersioningConfiguration(user, codeSystemToVersion.getShortName(), versionId, description, effectiveTime, force))
 						)
 					)
 				).execute(context);
@@ -163,16 +176,24 @@ final class CodeSystemVersionCreateRequest implements Request<ServiceProvider, B
 		}
 	}
 	
+	private boolean deleteBranch(ServiceProvider context, String path, String repositoryId) {
+		return RepositoryRequests.branching()
+				.prepareDelete(path)
+				.build(repositoryId)
+				.execute(context.service(IEventBus.class))
+				.getSync(1, TimeUnit.MINUTES);
+	}
+	
 	private Map<String, CodeSystem> fetchAllCodeSystems(ServiceProvider context) {
 		final RepositoryManager repositoryManager = context.service(RepositoryManager.class);
 		final Collection<Repository> repositories = repositoryManager.repositories();
 		return repositories.stream()
-			.map(repository -> fetchCodeSystem(context, repository.id()))
+			.map(repository -> fetchCodeSystems(context, repository.id()))
 			.flatMap(Collection::stream)
 			.collect(Collectors.toMap(CodeSystem::getShortName, Function.identity()));
 	}
 	
-	private List<CodeSystem> fetchCodeSystem(ServiceProvider context, String repositoryId) {
+	private List<CodeSystem> fetchCodeSystems(ServiceProvider context, String repositoryId) {
 		return CodeSystemRequests.prepareSearchCodeSystem()
 			.all()
 			.build(repositoryId)
@@ -181,31 +202,44 @@ final class CodeSystemVersionCreateRequest implements Request<ServiceProvider, B
 			.getItems();
 	}
 	
-	private void validateEffectiveTime(ServiceProvider context, CodeSystem codeSystem) {
+	private void validateVersion(ServiceProvider context, CodeSystem codeSystem) {
 		if (!context.service(TerminologyRegistry.class).getTerminology(codeSystem.getTerminologyId()).isEffectiveTimeSupported()) {
 			return;
 		}
 
-		Instant mostRecentVersionEffectiveTime = getMostRecentVersionEffectiveDateTime(context, codeSystem);
-		Instant requestEffectiveTime = effectiveTimeDate.toInstant();
+		Optional<CodeSystemVersion> mostRecentVersion = getMostRecentVersion(context, codeSystem);
 
-		if (!requestEffectiveTime.isAfter(mostRecentVersionEffectiveTime)) {
-			throw new BadRequestException("The specified '%s' effective time is invalid. Date should be after '%s'.", requestEffectiveTime, mostRecentVersionEffectiveTime);
-		}
+		mostRecentVersion.ifPresent(mrv -> {
+			LocalDate mostRecentVersionEffectiveTime = mostRecentVersion.map(CodeSystemVersion::getEffectiveTime).orElse(LocalDate.EPOCH);
+
+			if (force) {
+				if (!Objects.equals(versionId, mrv.getVersion())) {
+					throw new BadRequestException("Force creating version requires the same versionId ('%s') to be used", versionId);
+				}
+				
+				// force recreating an existing version should use the same or later effective date value, allow same here
+				if (effectiveTime.equals(mostRecentVersionEffectiveTime)) {
+					return;
+				}
+			}
+			
+			if (!effectiveTime.isAfter(mostRecentVersionEffectiveTime)) {
+				throw new BadRequestException("The specified '%s' effective time is invalid. Date should be after '%s'.", effectiveTime, mostRecentVersionEffectiveTime);
+			}
+		});
+		
 	}
 	
-	private Instant getMostRecentVersionEffectiveDateTime(ServiceProvider context, CodeSystem codeSystem) {
+	private Optional<CodeSystemVersion> getMostRecentVersion(ServiceProvider context, CodeSystem codeSystem) {
 		return CodeSystemRequests.prepareSearchCodeSystemVersion()
-			.all()
+			.one()
 			.filterByCodeSystemShortName(codeSystem.getShortName())
+			.sortBy(SortField.descending(CodeSystemVersionEntry.Fields.EFFECTIVE_DATE))
 			.build(codeSystem.getRepositoryId())
 			.execute(context.service(IEventBus.class))
 			.getSync(1, TimeUnit.MINUTES)
 			.stream()
-			.max(CodeSystemVersionEntry.VERSION_EFFECTIVE_DATE_COMPARATOR)
-			.map(CodeSystemVersionEntry::getEffectiveDate)
-			.map(Instant::ofEpochMilli)
-			.orElse(Instant.EPOCH);
+			.findFirst();
 	}
 	
 	private void acquireLocks(ServiceProvider context, String user, Collection<CodeSystem> codeSystems) {