+++ snomed/com.b2international.snowowl.snomed.datastore/src/com/b2international/snowowl/snomed/core/store/SnomedComponentBuilder.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2011-2020 B2i Healthcare Pte Ltd, http://b2i.sg
+ * Copyright 2011-2021 B2i Healthcare Pte Ltd, http://b2i.sg
  * 
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -15,7 +15,7 @@
  */
 package com.b2international.snowowl.snomed.core.store;
 
-import java.util.Date;
+import java.time.LocalDate;
 
 import javax.annotation.OverridingMethodsMustInvokeSuper;
 
@@ -36,7 +36,7 @@ public abstract class SnomedComponentBuilder<B extends SnomedComponentBuilder<B,
 
 	private String id;
 	private boolean active = true;
-	private Date effectiveTime;
+	private LocalDate effectiveTime;
 	private String moduleId;
 
 	/**
@@ -78,7 +78,7 @@ public abstract class SnomedComponentBuilder<B extends SnomedComponentBuilder<B,
 	 * @param effectiveTime
 	 * @return
 	 */
-	public final B withEffectiveTime(Date effectiveTime) {
+	public final B withEffectiveTime(LocalDate effectiveTime) {
 		this.effectiveTime = effectiveTime;
 		return getSelf();
 	}
@@ -104,7 +104,7 @@ public abstract class SnomedComponentBuilder<B extends SnomedComponentBuilder<B,
 			component.effectiveTime(EffectiveTimes.UNSET_EFFECTIVE_TIME);
 			component.released(false);
 		} else {
-			component.effectiveTime(effectiveTime.getTime());
+			component.effectiveTime(EffectiveTimes.getEffectiveTime(effectiveTime));
 			component.released(true);
 		}
 	}